simplified device init removing RemoteDevice class
