+++ carina-webdriver/src/main/java/com/qaprosoft/carina/core/foundation/webdriver/core/factory/impl/MobileFactory.java
@@ -19,7 +19,6 @@ import java.lang.invoke.MethodHandles;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.time.Duration;
-import java.util.Map;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -29,7 +28,6 @@ import org.openqa.selenium.remote.RemoteWebDriver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.qaprosoft.carina.commons.models.RemoteDevice;
 import com.qaprosoft.carina.core.foundation.commons.SpecialKeywords;
 import com.qaprosoft.carina.core.foundation.utils.Configuration;
 import com.qaprosoft.carina.core.foundation.utils.Configuration.Parameter;
@@ -178,11 +176,8 @@ public class MobileFactory extends AbstractFactory {
         } catch (MalformedURLException e) {
             throw new RuntimeException("Malformed selenium URL!", e);
         } catch (Exception e) {
-            Map<String, Object> capabilitiesMap = capabilities.asMap();
-            LOGGER.debug("Driver hasn't been created with capabilities: ".concat(capabilitiesMap.toString()));
-
             Device device = IDriverPool.nullDevice;
-            if (R.CONFIG.getBoolean("capabilities.STF_ENABLED")) {
+            if (R.CONFIG.getBoolean(SpecialKeywords.CAPABILITIES + "." + SpecialKeywords.STF_ENABLED)) {
                 LOGGER.debug("STF is enabled. Debug info will be extracted from the exception.");
                 if (e != null) {
                     String debugInfo = getDebugInfo(e.getMessage());
@@ -195,27 +190,15 @@ public class MobileFactory extends AbstractFactory {
                     }
                 }
             } else {
-                device = new Device(getDeviceInfo(capabilitiesMap));
+                device = new Device(capabilities);
             }
             IDriverPool.registerDevice(device);
             throw e;
         }
 
-        Device device = IDriverPool.getNullDevice();
-        if (device.isNull()) {
-            RemoteDevice remoteDevice = getDeviceInfo(driver);
-            // 3rd party solutions like browserstack or saucelabs return not
-            // null
-            if (remoteDevice != null && remoteDevice.getName() != null) {
-                device = new Device(remoteDevice);
-            } else if (driver != null) {
-                device = new Device(driver.getCapabilities());
-            }
-
-            IDriverPool.registerDevice(device);
-        }
-        // will be performed just in case uninstall_related_apps flag marked as
-        // true
+        Device device = new Device(driver.getCapabilities());
+        IDriverPool.registerDevice(device);
+        // will be performed just in case uninstall_related_apps flag marked as true
         device.uninstallRelatedApps();
 
         return driver;
@@ -223,69 +206,6 @@ public class MobileFactory extends AbstractFactory {
 
     private DesiredCapabilities getCapabilities(String name) {
         return new MobileCapabilies().getCapability(name);
-    }
-
-    /**
-     * Returns device information from Grid Hub using STF service.
-     * 
-     * @param drv
-     *            - driver
-     * @return remote device information
-     */
-    @SuppressWarnings("unchecked")
-    private RemoteDevice getDeviceInfo(RemoteWebDriver drv) {
-        return getDeviceInfo((Map<String, Object>) drv.getCapabilities().getCapability(SpecialKeywords.SLOT_CAPABILITIES));
-    }
-
-    /**
-     * Returns device information from Grid Hub using STF service.
-     * 
-     * @param cap
-     *            - capabilities
-     * @return remote device information
-     */
-    private RemoteDevice getDeviceInfo(Map<String, Object> cap) {
-        RemoteDevice remoteDevice = new RemoteDevice();
-        try {
-
-            if (cap != null && cap.containsKey("udid")) {
-
-                // restore device information from custom slotCapabilities map
-                /*
-                 * {deviceType=Phone, proxy_port=9000,
-                 * server:CONFIG_UUID=24130dde-59d4-4310-95ba-6f57b9d265c3,
-                 * seleniumProtocol=WebDriver, adb_port=5038,
-                 * vnc=wss://stage.qaprosoft.com:7410/websockify,
-                 * deviceName=Nokia_6_1, version=8.1.0, platform=ANDROID,
-                 * platformVersion=8.1.0, automationName=uiautomator2,
-                 * browserName=Nokia_6_1, maxInstances=1, platformName=ANDROID,
-                 * udid=PL2GAR9822804910}
-                 */
-
-                // TODO: remove code duplicates with carina-grid DeviceInfo
-                remoteDevice.setName((String) cap.get("deviceName"));
-                remoteDevice.setOs((String) cap.get("platformName"));
-                remoteDevice.setOsVersion((String) cap.get("platformVersion"));
-                remoteDevice.setType((String) cap.get("deviceType"));
-                remoteDevice.setUdid((String) cap.get("udid"));
-                if (cap.containsKey("vnc")) {
-                    remoteDevice.setVnc((String) cap.get("vnc"));
-                }
-                if (cap.containsKey(Parameter.PROXY_PORT.getKey())) {
-                    remoteDevice.setProxyPort(String.valueOf(cap.get(Parameter.PROXY_PORT.getKey())));
-                }
-
-                if (cap.containsKey("remoteURL")) {
-                    remoteDevice.setRemoteURL(String.valueOf(cap.get("remoteURL")));
-                }
-
-                remoteDevice.setCapabilities(new DesiredCapabilities(cap));
-            }
-
-        } catch (Exception e) {
-            LOGGER.error("Unable to get device info!", e);
-        }
-        return remoteDevice;
     }
 
     /**
